// src/lib/database/types.ts
export type Database = {
    public: {
        Tables: {
            alarm: {
                Row: {
                    created_at: string;
                    id: string;
                    last_notification_id: string | null;
                    next_alarm_time: string | null;
                    page_id: string;
                    processed_at: string | null;
                    sent_count: number | null;
                    updated_at: string;
                    user_id: string;
                };
                Insert: {
                    created_at?: string;
                    id: string;
                    last_notification_id?: string | null;
                    next_alarm_time?: string | null;
                    page_id: string;
                    processed_at?: string | null;
                    sent_count?: number | null;
                    updated_at?: string;
                    user_id?: string;
                };
                Update: {
                    created_at?: string;
                    id?: string;
                    last_notification_id?: string | null;
                    next_alarm_time?: string | null;
                    page_id?: string;
                    processed_at?: string | null;
                    sent_count?: number | null;
                    updated_at?: string;
                    user_id?: string;
                };
                Relationships: [
                    {
                        foreignKeyName: 'alarm_page_id_fkey';
                        columns: ['page_id'];
                        isOneToOne: true;
                        referencedRelation: 'page';
                        referencedColumns: ['id'];
                    },
                ];
            };
// ...
            folder: {
                Row: {
                    created_at: string;
                    description: string | null;
                    id: string;
                    last_page_added_at: string | null;
                    name: string;
                    page_count: number;
                    thumbnail_url: string | null;
                    updated_at: string;
                    user_id: string;
                };
                Insert: {
                    created_at?: string;
                    description?: string | null;
                    id: string;
                    last_page_added_at?: string | null;
                    name: string;
                    page_count?: number;
                    thumbnail_url?: string | null;
                    updated_at?: string;
                    user_id?: string;
                };
                Update: {
                    created_at?: string;
                    description?: string | null;
                    id?: string;
                    last_page_added_at?: string | null;
                    name?: string;
                    page_count?: number;
                    thumbnail_url?: string | null;
                    updated_at?: string;
                    user_id?: string;
                };
                Relationships: [];
            };
// ...
            page: {
                Row: {
                    body: string;
                    child_count: number | null;
                    created_at: string;
                    folder_id: string | null;
                    id: string;
                    img_url: string | null;
                    is_public: boolean | null;
                    last_embedded_at: string | null;
                    last_viewed_at: string | null;
                    length: number | null;
                    parent_count: number | null;
                    title: string;
                    type: Database['public']['Enums']['page_type'];
                    updated_at: string | null;
                    user_id: string;
                };
                Insert: {
                    body: string;
                    child_count?: number | null;
                    created_at?: string;
                    folder_id?: string | null;
                    id: string;
                    img_url?: string | null;
                    is_public?: boolean | null;
                    last_embedded_at?: string | null;
                    last_viewed_at?: string | null;
                    length?: number | null;
                    parent_count?: number | null;
                    title: string;
                    type?: Database['public']['Enums']['page_type'];
                    updated_at?: string | null;
                    user_id?: string;
                };
                Update: {
                    body?: string;
                    child_count?: number | null;
                    created_at?: string;
                    folder_id?: string | null;
                    id?: string;
                    img_url?: string | null;
                    is_public?: boolean | null;
                    last_embedded_at?: string | null;
                    last_viewed_at?: string | null;
                    length?: number | null;
                    parent_count?: number | null;
                    title?: string;
                    type?: Database['public']['Enums']['page_type'];
                    updated_at?: string | null;
                    user_id?: string;
                };
                Relationships: [
                    {
                        foreignKeyName: 'page_folder_id_fkey';
                        columns: ['folder_id'];
                        isOneToOne: false;
                        referencedRelation: 'folder';
                        referencedColumns: ['id'];
                    },
                ];
            };
// ...
        };
// ...
    };
};

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, 'public'>];
// ...
export type Enums<
    DefaultSchemaEnumNameOrOptions extends
        | keyof DefaultSchema['Enums']
        | { schema: keyof DatabaseWithoutInternals },
    EnumName extends DefaultSchemaEnumNameOrOptions extends {
        schema: keyof DatabaseWithoutInternals;
    }
        ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
        : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
}
    ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
    : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
      ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
      : never;